requirements.txt
django~=3.2
scrapy~=2.5.0
itemadapter~=0.2.0
django-mptt~=0.12.0